Renamed files and organised into folders
