+++ src/js/dataView/dataset/ImageIcon.js
